% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_module_preview_mobile.R
\name{mod_module_preview_mobile_ui}
\alias{mod_module_preview_mobile_ui}
\title{#' module_preview_mobile UI Function
#'
#' @description A shiny Module.
#'
#' @param id,input,output,session Internal parameters for {shiny}.
#'
#' @noRd
#'
#' @importFrom shiny NS tagList div h4
#' @importFrom golem get_golem_options
#' @importFrom RSQLite SQLite
#'}
\usage{
mod_module_preview_mobile_ui(id)
}
\description{
#' module_preview_mobile UI Function
#'
#' @description A shiny Module.
#'
#' @param id,input,output,session Internal parameters for {shiny}.
#'
#' @noRd
#'
#' @importFrom shiny NS tagList div h4
#' @importFrom golem get_golem_options
#' @importFrom RSQLite SQLite
#'
}
